fix: prompter file
